Add `src/directives` folder (vue directives refactor)
- create `src/directives` folder
- remove `src/app/gui/vue/directives/v-checked.js`
- remove `src/app/gui/vue/directives/v-disabled.js`
- remove `src/app/gui/vue/directives/v-select2.js`
- deprecate notice for `src/app/gui/vue/vue.directives.js`
- consolidate directives definitions within a single folder
