+++ src/directives/v-checked.js
@@ -1,3 +1,6 @@
+/**
+ * ORIGINAL SOURCE: src/app/gui/vue/directives/v-checked.js@v3.6
+ */
 export default (el, binding) => {
   binding.value ? el.setAttribute('checked','checked') : el.removeAttribute('checked');
 };
