+++ src/directives/v-select2.js
@@ -1,3 +1,6 @@
+/**
+ * ORIGINAL SOURCE: src/app/gui/vue/directives/v-select2.js@v3.6
+ */
 export default {
   inserted(el, binding, vnode){
     const { templateResult, templateSelection, multiple=false, search=true, select2_value} = vnode.data.attrs || {};
